requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.42.0
django-autoslug==1.9.9
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
